+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 + revamp apc sprites
 + revamp ifv sprites
 + revamp spg sprites
-+ added bomber
++ revamp helicopter sprites
++ added bomber, fighter
 + attempted nerf of spg and helicopter
 + for the love of god, correct version number
 + made vehicle cheaper